requirements.txt
fastapi
psycopg2
sqlalchemy
pandas
streamlit
streamlit_folium
altair
requests
datetime
geopy
python-dotenv